AddCase{Form,Input}: restructure into own dirs
