--- src/components/cases/AddCase/index.js
@@ -1,2 +1,0 @@
-export { default as AddCaseForm } from './Form';
-export { default as AddCaseInput } from './Input';
