+++ src/components/cases/AddCaseForm/AddCaseForm.test.jsx
@@ -4,19 +4,19 @@ import { Platform, View } from 'react-native';
 import renderer from 'react-test-renderer';
 import { render, fireEvent } from 'react-native-testing-library';
 
-import AddCaseInput from './Input';
+import AddCaseForm from './AddCaseForm';
 
 /**********************************************************/
 
-describe('AddCaseInput component', () => {
+describe('AddCaseForm component', () => {
 
     test('renders', () => {
-        const tree = renderer.create(<AddCaseInput/>).toJSON();
+        const tree = renderer.create(<AddCaseForm/>).toJSON();
         expect(tree.children).toHaveLength(1);
     });
 
     test('matches snapshot', () => {
-        const tree = renderer.create(<AddCaseInput/>).toJSON();
+        const tree = renderer.create(<AddCaseForm/>).toJSON();
         expect(tree).toMatchSnapshot();
     });
