+++ src/components/cases/AddCaseForm/index.js
@@ -1,0 +1,2 @@
+import AddCaseForm from './AddCaseForm';
+export default AddCaseForm;
