+++ src/components/cases/AddCaseInput/AddCaseInput.test.jsx
@@ -4,19 +4,19 @@ import { Platform, View } from 'react-native';
 import renderer from 'react-test-renderer';
 import { render, fireEvent } from 'react-native-testing-library';
 
-import AddCaseForm from './Form';
+import AddCaseInput from './AddCaseInput';
 
 /**********************************************************/
 
-describe('AddCaseForm component', () => {
+describe('AddCaseInput component', () => {
 
     test('renders', () => {
-        const tree = renderer.create(<AddCaseForm/>).toJSON();
+        const tree = renderer.create(<AddCaseInput/>).toJSON();
         expect(tree.children).toHaveLength(1);
     });
 
     test('matches snapshot', () => {
-        const tree = renderer.create(<AddCaseForm/>).toJSON();
+        const tree = renderer.create(<AddCaseInput/>).toJSON();
         expect(tree).toMatchSnapshot();
     });
