+++ src/components/cases/AddCaseInput/index.js
@@ -1,0 +1,2 @@
+import AddCaseInput from './AddCaseInput';
+export default AddCaseInput;
